fix: handle submittion date saving and viewing
